Include default format of JSON for the api routes.
The server will respond with JSON to a GET request coming to /students. You don't have to specify /students.json. However if the request goes to /students.html then rails with throw a 'UnknownFormat' excpetion and respond with a '406 Not Acceptable'. There's probably a way to handle this explicitly, however I'm not sure how to do this. I'm assuming here that we have no use for this data in html format.